feat : 테스트 계정에 대한 동작 추가

+++ src/main/java/com/app/guttokback/user/application/service/UserService.java
@@ -80,9 +80,15 @@ public class UserService {
     }
 
     @Transactional
-    public void userDelete(String email) {
+    public boolean userDelete(String email) {
         User user = findByUserEmail(email);
-        userRepository.delete(user);
+        if(user.getAuthorities().contains(Roles.ROLE_USER)) {
+            userRepository.delete(user);
+            return true;
+        } else {
+            return false;
+        }
+
     }
 
     public UserDetailInfo userDetail(String email) {
